# src/App/backend/test/Altinn.App.Api.Tests/OpenApi/OpenApiSpecChangeDetection.SaveCustomOpenApiSpec.verified.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Altinn 3 App API for Endring av navn (RF-1453)",
...
    "contact": {
      "name": "Digitaliseringsdirektoratet (altinn)",
      "url": "https://digdir-samarbeid.slack.com"
    },
    "version": ""
  },
  "servers": [
    {
      "url": "http://local.altinn.cloud",
      "description": "Local development server"
    },
    {
      "url": "https://tdd.apps.tt02.altinn.no",
      "description": "TT02 server"
    },
    {
      "url": "https://tdd.apps.altinn.no",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/tdd/contributer-restriction/instances/{instanceOwnerPartyId}/{instanceGuid}": {
      "summary": "Operations for instance",
      "description": "CRUD operations for a specific instance",
      "get": {
        "tags": [
          "Instances"
        ],
        "summary": "Get instance",
        "description": "Get a specific instance",
        "operationId": "GetInstance",
        "responses": {
          "200": {
            "description": "Instance found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Instance"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ProblemDetails"
          },
          "401": {
            "$ref": "#/components/responses/ProblemDetails"
          },
          "403": {
            "$ref": "#/components/responses/ProblemDetails"
          },
          "404": {
            "$ref": "#/components/responses/ProblemDetails"
          },
          "500": {
            "$ref": "#/components/responses/ProblemDetails"
          }
        }
      },
      "delete": {
        "tags": [
          "Instances"
        ],
        "summary": "Delete instance",
        "description": "Delete a specific instance",
        "responses": {
          "204": {
            "description": "Instance deleted"
          },
          "400": {
            "$ref": "#/components/responses/ProblemDetails"
          },
          "401": {
            "$ref": "#/components/responses/ProblemDetails"
          },
          "403": {
            "$ref": "#/components/responses/ProblemDetails"
          },
          "404": {
            "$ref": "#/components/responses/ProblemDetails"
          },
          "500": {
            "$ref": "#/components/responses/ProblemDetails"
          }
        }
      },
      "parameters": [
        {
          "$ref": "#/components/parameters/instanceOwnerPartyId"
        },
        {
          "$ref": "#/components/parameters/instanceGuid"
        },
        {
          "$ref": "#/components/parameters/language"
        }
      ]
    },
...
    "/tdd/contributer-restriction/instances/{instanceOwnerPartyId}/{instanceGuid}/data/{dataGuid}": {
      "summary": "Delete data element",
      "delete": {
        "tags": [
          "Instances"
        ],
        "summary": "Delete data element",
        "description": "Delete data for a specific data element",
        "responses": {
          "204": {
            "description": "Data deleted"
          },
          "404": {
            "description": "Data not found"
          }
        }
      },
      "parameters": [
        {
          "$ref": "#/components/parameters/instanceOwnerPartyId"
        },
        {
          "$ref": "#/components/parameters/instanceGuid"
        },
        {
          "$ref": "#/components/parameters/dataGuid"
        },
        {
          "$ref": "#/components/parameters/language"
        }
      ]
    },
...
  },
  "components": {
    "schemas": {
      "InstanceOwner": {
        "type": "object",
        "properties": {
          "partyId": {
            "type": "string",
            "nullable": true
          },
          "personNumber": {
            "type": "string",
            "nullable": true
          },
          "organisationNumber": {
            "type": "string",
            "nullable": true
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "externalIdentifier": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ResourceLinks": {
        "type": "object",
        "properties": {
          "apps": {
            "type": "string",
            "nullable": true
          },
          "platform": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ValidationStatus": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "canCompleteTask": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "ProcessElementInfo": {
        "type": "object",
        "properties": {
          "flow": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "started": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "elementId": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "altinnTaskType": {
            "type": "string",
            "nullable": true
          },
          "ended": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "validated": {
            "$ref": "#/components/schemas/ValidationStatus"
          },
          "flowType": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProcessState": {
        "type": "object",
        "properties": {
          "started": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "startEvent": {
            "type": "string",
            "nullable": true
          },
          "currentTask": {
            "$ref": "#/components/schemas/ProcessElementInfo"
          },
          "ended": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "endEvent": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ReadStatus": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "Substatus": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "InstanceStatus": {
        "type": "object",
        "properties": {
          "isArchived": {
            "type": "boolean"
          },
          "archived": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isSoftDeleted": {
            "type": "boolean"
          },
          "softDeleted": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isHardDeleted": {
            "type": "boolean"
          },
          "hardDeleted": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "readStatus": {
            "$ref": "#/components/schemas/ReadStatus"
          },
          "substatus": {
            "$ref": "#/components/schemas/Substatus"
          }
        },
        "additionalProperties": false
      },
      "CompleteConfirmation": {
        "type": "object",
        "properties": {
          "stakeholderId": {
            "type": "string",
            "nullable": true
          },
          "confirmedOn": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "KeyValueEntry": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DeleteStatus": {
        "type": "object",
        "properties": {
          "isHardDeleted": {
            "type": "boolean"
          },
          "hardDeleted": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FileScanResult": {
        "enum": [
          "NotApplicable",
          "Pending",
          "Clean",
          "Infected"
        ],
        "type": "string"
      },
      "RelationType": {
        "enum": [
          "GeneratedFrom"
        ],
        "type": "string"
      },
      "ReferenceType": {
        "enum": [
          "DataElement",
          "Task"
        ],
        "type": "string"
      },
      "Reference": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "nullable": true
          },
          "relation": {
            "$ref": "#/components/schemas/RelationType"
          },
          "valueType": {
            "$ref": "#/components/schemas/ReferenceType"
          }
        },
        "additionalProperties": false
      },
      "DataElement": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "created": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "lastChanged": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastChangedBy": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "instanceGuid": {
            "type": "string",
            "nullable": true
          },
          "dataType": {
            "type": "string",
            "nullable": true
          },
          "filename": {
            "type": "string",
            "nullable": true
          },
          "contentType": {
            "type": "string",
            "nullable": true
          },
          "blobStoragePath": {
            "type": "string",
            "nullable": true
          },
          "selfLinks": {
            "$ref": "#/components/schemas/ResourceLinks"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "contentHash": {
            "type": "string",
            "nullable": true
          },
          "contentEtag": {
            "type": "string",
            "nullable": true
          },
          "locked": {
            "type": "boolean"
          },
          "refs": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "isRead": {
            "type": "boolean"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "userDefinedMetadata": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KeyValueEntry"
            },
            "nullable": true
          },
          "metadata": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KeyValueEntry"
            },
            "nullable": true
          },
          "deleteStatus": {
            "$ref": "#/components/schemas/DeleteStatus"
          },
          "fileScanResult": {
            "$ref": "#/components/schemas/FileScanResult"
          },
          "references": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Reference"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Instance": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "created": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "lastChanged": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastChangedBy": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "instanceOwner": {
            "$ref": "#/components/schemas/InstanceOwner"
          },
          "appId": {
            "type": "string",
            "nullable": true
          },
          "org": {
            "type": "string",
            "nullable": true
          },
          "selfLinks": {
            "$ref": "#/components/schemas/ResourceLinks"
          },
          "dueBefore": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "visibleAfter": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "process": {
            "$ref": "#/components/schemas/ProcessState"
          },
          "status": {
            "$ref": "#/components/schemas/InstanceStatus"
          },
          "completeConfirmations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompleteConfirmation"
            },
            "nullable": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataElement"
            },
            "nullable": true
          },
          "presentationTexts": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "nullable": true
            },
            "nullable": true
          },
          "dataValues": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "nullable": true
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "responses": {
      "ProblemDetails": {
        "description": "Problem details",
        "content": {
          "application/problem+json": {
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "nullable": true,
                  "example": "https://datatracker.ietf.org/doc/html/rfc6902/"
                },
                "title": {
                  "type": "string",
                  "nullable": true,
                  "example": "Error in data processing"
                },
                "status": {
                  "type": "integer",
                  "format": "int32",
                  "nullable": true,
                  "example": 400
                },
                "detail": {
                  "type": "string",
                  "nullable": true,
                  "example": "Actually useful description of the error"
                },
                "instance": {
                  "type": "string",
                  "nullable": true
                },
                "traceId": {
                  "type": "string",
                  "nullable": true
                }
              },
              "additionalProperties": {
                "title": "Additional properties",
                "description": "Additional properties can be added to the problem details"
              }
            }
          }
        }
      }
    },
    "parameters": {
      "instanceOwnerPartyId": {
        "name": "instanceOwnerPartyId",
        "in": "path",
        "description": "PartyId for the owner of the instance, this is Altinn's internal id for the organisation, person or self registered user. Might be the current user, or a party the user has rights to represent.",
        "required": true,
        "schema": {
          "type": "integer"
        }
      },
      "instanceGuid": {
        "name": "instanceGuid",
        "in": "path",
        "description": "The guid part of instance.Id",
        "required": true,
        "schema": {
          "type": "string",
          "format": "guid"
        }
      },
      "dataGuid": {
        "name": "dataGuid",
        "in": "path",
        "description": "Id of this data element that belongs to an instance",
        "required": true,
        "schema": {
          "type": "string",
          "format": "guid"
        }
      },
      "language": {
        "name": "language",
        "in": "query",
        "description": "Some apps make changes to the data models or validation based on the active language of the user",
        "schema": {
          "pattern": "\\w\\w",
          "type": "string"
        },
        "example": "nb"
      }
    },
    "securitySchemes": {
      "AltinnToken": {
        "type": "http",
        "description": "Get a token for [localtest](http://local.altinn.cloud/Home/Tokens)\nor by exchanging a Maskinporten token with the [token exchange endpoint](https://docs.altinn.studio/api/authentication/spec/#/Authentication/get_exchange__tokenProvider_)",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "AltinnToken": []
    }
  ],
...
}